requirements-a2a.txt
# Core ADK framework + A2A server
google-adk>=1.25.0
a2a-sdk[http-server]>=0.3.0

# HTTP client for FHIR tools
httpx>=0.28.0

# Environment variable loading
python-dotenv>=1.0.0

# ASGI server
uvicorn>=0.41.0

# PDF parsing for prior operative notes
pypdf>=4.0
